refactor(capabilities): rename module to tool_search.rs
