+++ crates/core/src/capabilities/tool_search.rs
